+++ bittide-instances/bittide-instances.cabal
@@ -178,6 +178,7 @@ library
     Bittide.Instances.Hitl.Driver.SwitchDemo
     Bittide.Instances.Hitl.Driver.VexRiscv
     Bittide.Instances.Hitl.Driver.VexRiscvTcp
+    Bittide.Instances.Hitl.Dut.SwitchDemo
     Bittide.Instances.Hitl.Ethernet
     Bittide.Instances.Hitl.FincFdec
     Bittide.Instances.Hitl.LinkConfiguration
