+++ bittide-instances/src/Bittide/Instances/Hitl/SwitchDemo.hs
@@ -2,7 +2,6 @@
 --
 -- SPDX-License-Identifier: Apache-2.0
 {-# LANGUAGE CPP #-}
-{-# LANGUAGE RecordWildCards #-}
 
 {- | Switch demo for a Bittide system. In concert with its driver file, this device under
 test should demonstrate the predictability of a Bittide system once it has achieved logical
@@ -14,28 +13,15 @@ on the topic.
 module Bittide.Instances.Hitl.SwitchDemo (
   switchDemoDut,
   switchDemoTest,
-  memoryMapCc,
-  memoryMapMu,
   tests,
 ) where
 
 import Clash.Explicit.Prelude
-import Clash.Prelude (HiddenClockResetEnable, withClockResetEnable)
+import Clash.Prelude (withClockResetEnable)
 
-import Bittide.Calendar (CalendarConfig (..), ValidEntry (..))
-import Bittide.CaptureUgn (captureUgn)
 import Bittide.ClockControl
-import Bittide.ClockControl.Callisto.Types (
-  CallistoResult (..),
-  Stability (..),
- )
-import Bittide.ClockControl.CallistoSw (
-  SwcccInternalBusses,
-  callistoSwClockControlC,
- )
+import Bittide.ClockControl.Callisto.Types (CallistoResult (..))
 import Bittide.ClockControl.Si539xSpi (ConfigState (Error, Finished), si539xSpi)
-import Bittide.Df (asciiDebugMux)
-import Bittide.DoubleBufferedRam
 import Bittide.ElasticBuffer (
   EbMode (Pass),
   Overflow,
@@ -58,200 +44,30 @@ import Bittide.Instances.Domains (
   GthRxS,
   GthTxS,
  )
-import Bittide.Instances.Hitl.Setup (
-  FpgaCount,
-  LinkCount,
-  allHwTargets,
-  channelNames,
-  clockPaths,
- )
-import Bittide.Jtag (jtagChain)
-import Bittide.ProcessingElement (
-  PeConfig (..),
-  PrefixWidth,
-  RemainingBusWidth,
-  processingElement,
- )
-import Bittide.SharedTypes (Byte, Bytes, withBittideByteOrder)
-import Bittide.Switch (switchC)
-import Bittide.SwitchDemoProcessingElement (SimplePeState (Idle), switchDemoPeWb)
+import Bittide.Instances.Hitl.Dut.SwitchDemo (circuitFnC)
+import Bittide.Instances.Hitl.Setup (LinkCount, allHwTargets, channelNames, clockPaths)
+import Bittide.SharedTypes (withBittideByteOrder)
+import Bittide.SwitchDemoProcessingElement (SimplePeState (Idle))
 import Bittide.Transceiver (transceiverPrbsN)
-import Bittide.Wishbone (
-  readDnaPortE2Wb,
-  timeWb,
-  uartBytes,
-  uartDf,
-  uartInterfaceWb,
-  whoAmIC,
- )
 
 import Clash.Annotations.TH (makeTopEntity)
-import Clash.Class.BitPackC (ByteOrder)
-import Clash.Cores.Xilinx.DcFifo (dcFifoDf)
 import Clash.Cores.Xilinx.Ila (Depth (..), IlaConfig (..), ila, ilaConfig)
-import Clash.Cores.Xilinx.Unisim.DnaPortE2 (simDna2)
 import Clash.Cores.Xilinx.VIO (vioProbe)
 import Clash.Cores.Xilinx.Xpm.Cdc (xpmCdcArraySingle, xpmCdcSingle)
 import Clash.Xilinx.ClockGen (clockWizardDifferential)
-import Data.Char (ord)
 import Protocols
-import Protocols.Extra
-import Protocols.MemoryMap (ConstBwd, MM, MemoryMap)
-import Protocols.Wishbone
 import System.FilePath ((</>))
-import VexRiscv (DumpVcd (..), Jtag, JtagIn (..), JtagOut (..))
+import VexRiscv (JtagIn (..), JtagOut (..))
 #ifdef SIM_BAUD_RATE
 import Clash.Cores.UART.Extra
 #endif
 
-import {-# SOURCE #-} qualified Bittide.Instances.Hitl.Driver.SwitchDemo as D
+import qualified Bittide.Instances.Hitl.Driver.SwitchDemo as D
 import qualified Bittide.Transceiver as Transceiver
 import qualified Clash.Cores.Xilinx.GTH as Gth
 import qualified Protocols.MemoryMap as MM
-import qualified Protocols.Vec as Vec
-
-#ifdef SIM_BAUD_RATE
-type Baud = MaxBaudRate Basic125
-#else
-type Baud = 921_600
-#endif
-
-baud :: SNat Baud
-baud = SNat
 
 type FifoSize = 5 -- = 2^5 = 32
-
-{- Internal busses:
-    - Instruction memory
-    - Data memory
-    - `timeWb`
--}
-type NmuInternalBusses = 3
-type NmuRemBusWidth nodeBusses = RemainingBusWidth (nodeBusses + NmuInternalBusses)
-
-data SimpleManagementConfig nodeBusses where
-  SimpleManagementConfig ::
-    (KnownNat nodeBusses) =>
-    { peConfig :: PeConfig (nodeBusses + NmuInternalBusses)
-    -- ^ Configuration for the internal 'processingElement'
-    , dumpVcd :: DumpVcd
-    -- ^ VCD dump configuration
-    } ->
-    SimpleManagementConfig nodeBusses
-
-simpleManagementUnitC ::
-  forall bitDom nodeBusses.
-  ( KnownDomain bitDom
-  , HiddenClockResetEnable bitDom
-  , ?busByteOrder :: ByteOrder
-  , ?regByteOrder :: ByteOrder
-  , 1 <= DomainPeriod bitDom
-  , KnownNat nodeBusses
-  , PrefixWidth (nodeBusses + NmuInternalBusses) <= 30
-  ) =>
-  SimpleManagementConfig nodeBusses ->
-  Circuit
-    (MM.ConstBwd MM.MM, (Jtag bitDom, CSignal bitDom (BitVector 64)))
-    ( CSignal bitDom (Unsigned 64)
-    , Vec
-        nodeBusses
-        ( MM.ConstBwd MM.MM
-        , Wishbone bitDom 'Standard (NmuRemBusWidth nodeBusses) (Bytes 4)
-        )
-    )
-simpleManagementUnitC (SimpleManagementConfig peConfig dumpVcd) =
-  circuit $ \(mm, (jtag, _linkIn)) -> do
-    peWbs <- processingElement dumpVcd peConfig -< (mm, jtag)
-    ([timeWbBus], nmuWbs) <- splitAtC d1 -< peWbs
-    localCounter <- timeWb -< timeWbBus
-
-    idC -< (localCounter, nmuWbs)
-
-{- FOURMOLU_DISABLE -} -- Fourmolu doesn't do well with tabular code
-calendarConfig :: CalendarConfig 26 (Vec 8 (Index 9))
-calendarConfig =
-  CalendarConfig
-    (SNat @LinkCount)
-    {- The '@12' is so that the generated Rust code works. At time of writing,
-    the generator makes two separate device-specific types for 'ValidEntry' since
-    they have differing repetition bit widths. To fix this, all tests are being
-    set to a width of 12.
-    -}
-    (SNat @12)
-
-    -- Active calendar. It will broadcast the PE (node 1) data to all links. Other
-    -- than that we cycle through the other nodes.
-    (      ValidEntry (2 :> repeat 1) nRepetitions
-        :> ValidEntry (3 :> repeat 1) nRepetitions
-        :> ValidEntry (4 :> repeat 1) nRepetitions
-        :> ValidEntry (5 :> repeat 1) nRepetitions
-        :> ValidEntry (6 :> repeat 1) nRepetitions
-        :> ValidEntry (7 :> repeat 1) nRepetitions
-        :> ValidEntry (8 :> repeat 1) nRepetitions
-        :> Nil
-    )
-
-    -- Don't care about inactive calendar:
-    (ValidEntry (repeat 0) 0 :> Nil)
-  where
-  -- We want enough time to read _number of FPGAs_ triplets
-  nRepetitions = numConvert (maxBound :: Index (FpgaCount * 3))
-{- FOURMOLU_ENABLE -}
-
-memoryMapCc :: MemoryMap
-memoryMapCc = fst memoryMaps
-
-memoryMapMu :: MemoryMap
-memoryMapMu = snd memoryMaps
-
-memoryMaps :: ("CC" ::: MemoryMap, "MU" ::: MemoryMap)
-memoryMaps = (ccMm, muMm)
- where
-  (SimOnly ccMm, SimOnly muMm, _, _, _, _, _, _, _, _, _, _, _, _, _, _) =
-    switchDemoDut
-      clockGen
-      resetGen
-      clockGen
-      (SimOnly (repeat 0))
-      0
-      0
-      0
-      (pure (JtagIn 0 0 0))
-      0
-
-muConfig :: SimpleManagementConfig 12
-muConfig =
-  SimpleManagementConfig
-    { peConfig =
-        PeConfig
-          { initI = Undefined @(Div (64 * 1024) 4)
-          , initD = Undefined @(Div (64 * 1024) 4)
-          , iBusTimeout = d0
-          , dBusTimeout = d0
-          , includeIlaWb = False
-          }
-    , dumpVcd = NoDumpVcd
-    }
-
-ccConfig ::
-  ( KnownNat n
-  , PrefixWidth (n + SwcccInternalBusses) <= 30
-  ) =>
-  PeConfig (n + SwcccInternalBusses)
-ccConfig =
-  PeConfig
-    { initI = Undefined @(Div (64 * 1024) 4)
-    , initD = Undefined @(Div (64 * 1024) 4)
-    , iBusTimeout = d0
-    , dBusTimeout = d0
-    , includeIlaWb = False
-    }
-
-ccLabel :: Vec 2 Byte
-ccLabel = fromIntegral (ord 'C') :> fromIntegral (ord 'C') :> Nil
-
-muLabel :: Vec 2 Byte
-muLabel = fromIntegral (ord 'M') :> fromIntegral (ord 'U') :> Nil
 
 {- | Reset logic:
 
@@ -471,147 +287,7 @@ switchDemoDut refClk refRst skyClk rxSims rxNs rxPs miso jtagIn syncIn =
   transceiversFailedAfterUp =
     sticky refClk refRst (isFalling refClk spiRst enableGen False handshakesDoneFree)
 
-  defaultBittideClkRstEn :: forall r. ((HiddenClockResetEnable Bittide) => r) -> r
-  defaultBittideClkRstEn = withClockResetEnable bittideClk handshakeRstTx enableGen
-  defaultRefClkRstEn :: forall r. ((HiddenClockResetEnable Basic125) => r) -> r
-  defaultRefClkRstEn = withClockResetEnable refClk handshakeRstFree enableGen
-
-  circuitFnC ::
-    ( ?busByteOrder :: ByteOrder
-    , ?regByteOrder :: ByteOrder
-    ) =>
-    Circuit
-      ( "CC" ::: ConstBwd MM
-      , "MU" ::: ConstBwd MM
-      , Jtag Bittide
-      , CSignal Bittide (BitVector 64)
-      , CSignal Bittide (BitVector LinkCount)
-      , Vec LinkCount (CSignal Bittide (Maybe (BitVector 64)))
-      )
-      ( CSignal Bittide (CallistoResult LinkCount)
-      , "TXS" ::: Vec LinkCount (CSignal Bittide (BitVector 64))
-      , "LOCAL_COUNTER" ::: CSignal Bittide (Unsigned 64)
-      , "PE_STATE" ::: CSignal Bittide (SimplePeState FpgaCount)
-      , "PE_IN" ::: CSignal Bittide (BitVector 64)
-      , "PE_OUT" ::: CSignal Bittide (BitVector 64)
-      , "CAL_ENTRY" ::: CSignal Bittide (Vec (LinkCount + 1) (Index (LinkCount + 2)))
-      , "UART_TX" ::: CSignal Basic125 Bit
-      , "SYNC_OUT" ::: CSignal Basic125 Bit
-      )
-  circuitFnC = circuit $ \(ccMM, muMM, jtag, linkIn, mask, Fwd rxs) -> do
-    [muJtag, ccJtag] <- jtagChain -< jtag
-
-    (muUartBytesBittide, _muUartStatus) <-
-      defaultBittideClkRstEn
-        $ uartInterfaceWb d16 d16 uartBytes
-        -< (muUartBus, Fwd (pure Nothing))
-
-    muUartBytes <-
-      dcFifoDf d5 bittideClk handshakeRstTx refClk handshakeRstFree
-        -< muUartBytesBittide
-
-    (Fwd lc, muWbAll) <-
-      defaultBittideClkRstEn (simpleManagementUnitC muConfig) -< (muMM, (muJtag, linkIn))
-
-    ( [ muWhoAmIWb
-        , (peWbMM, peWb)
-        , (switchWbMM, switchWb)
-        , dnaWb
-        , muUartBus
-        ]
-      , ugnData
-      ) <-
-      splitAtC SNat -< muWbAll
-
-    ugnRxs <-
-      defaultBittideClkRstEn $ Vec.vecCircuits (captureUgn lc <$> rxs) -< ugnData
-
-    rxLinks <- appendC -< ([Fwd peOut], ugnRxs)
-    (switchOut, calEntry) <-
-      defaultBittideClkRstEn $ switchC calendarConfig -< (switchWbMM, (rxLinks, switchWb))
-    ([Fwd peIn], txs) <- splitAtC SNat -< switchOut
-
-    (Fwd peOut, ps) <-
-      defaultBittideClkRstEn (switchDemoPeWb (SNat @FpgaCount))
-        -< (peWbMM, (Fwd lc, peWb, dna, Fwd peIn))
-
-    dna <- defaultBittideClkRstEn (readDnaPortE2Wb simDna2) -< dnaWb
-
-    defaultBittideClkRstEn (whoAmIC 0x746d_676d) -< muWhoAmIWb
-
-    (ccUartBytesBittide, _uartStatus) <-
-      defaultBittideClkRstEn
-        $ uartInterfaceWb d16 d16 uartBytes
-        -< (ccUartBus, Fwd (pure Nothing))
-
-    ccUartBytes <-
-      dcFifoDf d5 bittideClk handshakeRstTx refClk handshakeRstFree
-        -< ccUartBytesBittide
-
-    uartTxBytes <-
-      defaultRefClkRstEn
-        $ asciiDebugMux d1024 (ccLabel :> muLabel :> Nil)
-        -< [ccUartBytes, muUartBytes]
-    (_uartInBytes, uartTx) <- defaultRefClkRstEn $ uartDf baud -< (uartTxBytes, Fwd 0)
-
-    ( syncOut
-      , Fwd swCcOut0
-      , [ ccWhoAmIBus
-          , ccUartBus
-          , ccSampleMemoryBus
-          ]
-      ) <-
-      defaultBittideClkRstEn
-        $ callistoSwClockControlC
-          @LinkCount
-          refClk
-          refRst
-          transceivers.rxClocks
-          (unsafeFromActiveLow <$> transceivers.handshakesDone)
-          NoDumpVcd
-          ccConfig
-        -< (ccMM, (Fwd syncIn, ccJtag, mask, Fwd linksSuitableForCc))
-
-    defaultBittideClkRstEn
-      (wbStorage "SampleMemory")
-      (Undefined @36_000 @(BitVector 32))
-      -< ccSampleMemoryBus
-
-    defaultBittideClkRstEn (whoAmIC 0x6363_7773) -< ccWhoAmIBus
-
-    let swCcOut1 =
-          if clashSimulation
-            then
-              let
-                -- Should all clock control steps be run in simulation?
-                -- False means that clock control will always immediately be done.
-                simulateCc = False
-               in
-                if simulateCc
-                  then swCcOut0
-                  else
-                    pure
-                      $ CallistoResult
-                        { maybeSpeedChange = Nothing
-                        , stability = repeat (Stability{stable = True, settled = True})
-                        , allStable = True
-                        , allSettled = True
-                        }
-            else swCcOut0
-
-    idC
-      -< ( Fwd swCcOut1
-         , txs
-         , Fwd lc
-         , ps
-         , Fwd peIn
-         , Fwd peOut
-         , calEntry
-         , uartTx
-         , syncOut
-         )
-
-  ( (ccMm, muMm, jtagOut, _linkInBwd, _maskBwd, _insBwd)
+  ( (ccMm, muMm, jtagOut, _linkInBwd, _maskBwd, _l4ccBwd, _insBwd, _syncBwd)
     , ( callistoResult
         , switchDataOut
         , localCounter
@@ -625,14 +301,21 @@ switchDemoDut refClk refRst skyClk rxSims rxNs rxPs miso jtagIn syncIn =
     ) =
       withBittideByteOrder
         $ toSignals
-          circuitFnC
+          ( circuitFnC
+              (refClk, handshakeRstFree, enableGen)
+              (bittideClk, handshakeRstTx, enableGen)
+              transceivers.rxClocks
+              (unsafeFromActiveLow <$> transceivers.handshakesDone)
+          )
           (
             ( ()
             , ()
             , jtagIn
             , pure 0 -- link in
             , pure maxBound -- enable mask
+            , linksSuitableForCc
             , rxDatasEbs
+            , syncIn
             )
           ,
             ( pure ()
